Added TeamBitmask tables

+++ SF3Lib/Models/Structs/X1/Battle/XYCoord.cs
@@ -14,15 +14,15 @@ namespace SF3.Models.Structs.X1.Battle {
 
         [TableViewModelColumn(addressField: nameof(_xAddr), displayOrder: 0, minWidth: 60)]
         [BulkCopy]
-        public int X {
-            get => Data.GetWord(_xAddr);
+        public short X {
+            get => (short) Data.GetWord(_xAddr);
             set => Data.SetWord(_xAddr, value);
         }
 
         [TableViewModelColumn(addressField: nameof(_zAddr), displayOrder: 1, minWidth: 60)]
         [BulkCopy]
-        public int Z {
-            get => Data.GetWord(_zAddr);
+        public short Z {
+            get => (short) Data.GetWord(_zAddr);
             set => Data.SetWord(_zAddr, value);
         }
     }
